여러 파일 업로드 기능 추가

+++ src/main/webapp/uploadAction.jsp
@@ -1,3 +1,4 @@
+<%@page import="java.util.Enumeration"%>
 <%@page import="file.FileDAO"%>
 <%@page import="java.io.File"%>
 <%@page import="com.oreilly.servlet.multipart.DefaultFileRenamePolicy"%>
@@ -16,29 +17,59 @@
 		String directory = "D:/kdt_juri/jsp/upload";
 		int maxSize = 1024 * 1024 * 100; //100M까지만 가능
 		String encoding = "UTF-8";
+		
 		//동일한 이름의 파일을 저장할 때 저장할 파일의 이름을 바꾸는 정책을 사용(덮어씌우는 정책도 있다)
 		//이 예제에서는 모든 파일을 업로드하지만, 원래는 파일의 확장자를 제한하기 위해 시큐어코딩을 사용한다(보안성)
 		MultipartRequest multipartRequest
 		= new MultipartRequest(request, directory, maxSize, encoding,
 													new DefaultFileRenamePolicy());
 		
-		String fileName = multipartRequest.getOriginalFileName("file");
-		String fileRealName = multipartRequest.getFilesystemName("file");
+		Enumeration fileNames = multipartRequest.getFileNames();
 		
-		// 특정 파일만 불러 올 수 있고 그 외의 파일은 불러와도 바로 삭제
+		while(fileNames.hasMoreElements()){
+			String parameter = (String) fileNames.nextElement();
+			String fileName = multipartRequest.getOriginalFileName(parameter);
+		String fileRealName = multipartRequest.getFilesystemName(parameter);
+		
+		if(fileName == null) continue; // 파일을 선택하지 않은 항목에 대해서는 pass
+		// 특정 파일만 불러 올 수 있고 그 외에 파일은 불러와도 바로 삭제됨.
 		if(!fileName.endsWith(".gif") && !fileName.endsWith(".png") &&
-				!fileName.endsWith(".jpg") && !fileName.endsWith(".txt")){
-			
-			File file = new File(directory + "/" + fileRealName);
+			!fileName.endsWith(".jpg") && !fileName.endsWith(".txt")){
+			File file = new File(directory + fileRealName);
 			file.delete();
-			out.write("업로드 할 수 없는 확장자입니다.");
-			
+			out.write("업로드할 수 없는 확장자입니다.");
 		}else{
 			new FileDAO().upload(fileName, fileRealName);
-			out.write("파일명: "+fileName + "<br>");
-			out.write("실제파일명: "+fileRealName + "<br>");
+			out.write("파일명: " + fileName + "<br>");
+			out.write("실제파일명: " + fileRealName + "<br>");
+		}			
+	}
+		/*
+		
+		
+		
+		Enumeration fileNames = multipartRequest.getFileNames();
+		while(fileNames.hasMoreElements()){
+			String parameter = (String) fileNames.nextElement();
+			String fileName = multipartRequest.getOriginalFileName(parameter);
+		String fileRealName = multipartRequest.getFilesystemName(parameter);
+			
+			
+		if(fileName == null) continue; // 파일을 선택하지 않은 항목에 대해서는 pass
+			// 특정 파일만 불러 올 수 있고 그 외에 파일은 불러와도 바로 삭제됨.
+			if(!fileName.endsWith(".gif") && !fileName.endsWith(".png") &&
+				!fileName.endsWith(".jpg") && !fileName.endsWith(".txt")){
+				File file = new File(directory + fileRealName);
+				file.delete();
+				out.write("업로드할 수 없는 확장자입니다.");
+			}else{
+				new FileDAO().upload(fileName, fileRealName);
+				out.write("파일명: " + fileName + "<br>");
+				out.write("실제파일명: " + fileRealName + "<br>");
+			}			
 		}
 		
+		*/
 	%>
 </body>
 </html>
